refactor(*): remove reflection to log user auditor
