+++ src/main/java/sleppynavigators/studyupbackend/domain/group/Group.java
@@ -3,7 +3,6 @@ package sleppynavigators.studyupbackend.domain.group;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -15,9 +14,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.SoftDelete;
-import sleppynavigators.studyupbackend.domain.common.TimeAuditBaseEntity;
-import sleppynavigators.studyupbackend.domain.common.UserAuditAttribute;
-import sleppynavigators.studyupbackend.infrastructure.common.attribute.listener.UserAuditListener;
+import sleppynavigators.studyupbackend.domain.common.UserAndTimeAuditBaseEntity;
 import sleppynavigators.studyupbackend.domain.challenge.Challenge;
 import sleppynavigators.studyupbackend.domain.group.vo.GroupDetail;
 import sleppynavigators.studyupbackend.domain.user.User;
@@ -26,16 +23,12 @@ import sleppynavigators.studyupbackend.exception.business.ActionRequiredBeforeEx
 @SoftDelete
 @Entity(name = "`groups`")
 @Getter
-@EntityListeners(UserAuditListener.class)
 @NoArgsConstructor(access = lombok.AccessLevel.PROTECTED)
-public class Group extends TimeAuditBaseEntity {
+public class Group extends UserAndTimeAuditBaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-    @Embedded
-    private UserAuditAttribute userAuditAttribute;
 
     @Embedded
     private GroupDetail groupDetail;
