+++ src/main/java/sleppynavigators/studyupbackend/domain/group/invitation/GroupInvitation.java
@@ -1,9 +1,7 @@
 package sleppynavigators.studyupbackend.domain.group.invitation;
 
 import jakarta.persistence.Column;
-import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -14,23 +12,17 @@ import java.util.UUID;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import sleppynavigators.studyupbackend.domain.common.TimeAuditBaseEntity;
-import sleppynavigators.studyupbackend.domain.common.UserAuditAttribute;
-import sleppynavigators.studyupbackend.infrastructure.common.attribute.listener.UserAuditListener;
+import sleppynavigators.studyupbackend.domain.common.UserAndTimeAuditBaseEntity;
 import sleppynavigators.studyupbackend.domain.group.Group;
 
 @Entity(name = "group_invitations")
 @Getter
-@EntityListeners(UserAuditListener.class)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class GroupInvitation extends TimeAuditBaseEntity {
+public class GroupInvitation extends UserAndTimeAuditBaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
-    @Embedded
-    private UserAuditAttribute userAuditAttribute;
 
     @OneToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "group_id", nullable = false, updatable = false)
