+++ src/main/java/sleppynavigators/studyupbackend/infrastructure/common/config/JpaConfig.java
@@ -1,9 +1,16 @@
 package sleppynavigators.studyupbackend.infrastructure.common.config;
 
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.data.domain.AuditorAware;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @Configuration
 @EnableJpaAuditing
 public class JpaConfig {
+
+    @Bean
+    public AuditorAware<?> auditorProvider() {
+        return new UserIdAuditorAware();
+    }
 }
