+++ src/main/java/sleppynavigators/studyupbackend/infrastructure/common/config/UserIdAuditorAware.java
@@ -1,0 +1,24 @@
+package sleppynavigators.studyupbackend.infrastructure.common.config;
+
+import java.util.Optional;
+import org.jetbrains.annotations.NotNull;
+import org.springframework.data.domain.AuditorAware;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
+import sleppynavigators.studyupbackend.presentation.authentication.filter.UserPrincipal;
+
+public class UserIdAuditorAware implements AuditorAware<Long> {
+
+    @NotNull
+    @Override
+    public Optional<Long> getCurrentAuditor() {
+        return Optional.ofNullable(SecurityContextHolder.getContext())
+                .map(SecurityContext::getAuthentication)
+                .filter(Authentication::isAuthenticated)
+                .map(Authentication::getPrincipal)
+                .filter(UserPrincipal.class::isInstance)
+                .map(UserPrincipal.class::cast)
+                .map(UserPrincipal::userId);
+    }
+}
